scripts: Create an itlb monitor with metrices
To monitor a system for ITLB pressure one can use perf. Brendan Gregg already
wrote a simple script for this task. [1]

However, Brendans original tlbstat script focuses purely on ITLB and DTLB
stats. In addition, we would also like to see the recommended metrics from
Intel [2] and save the values separately in a file.

Therefore, implement a ITLB monitor script with the recommended intel
performance metrices and save-to-file feature.

[1] https://github.com/brendangregg/pmc-cloud-tools/blob/master/tlbstat
[2] https://www.intel.com/content/www/us/en/developer/articles/technical/runtime-performance-optimization-blueprint-intel-architecture-optimization-with-large-code.html
Signed-off-by: Anton Gres <dalo@linutronix.de>