workload: Create network workload
Because not all subsystems are prioritized equally in the kernel, more
interesting subsystems should be stress tested individually to find out their
hot caller-callee relationships. However, we don't want to focus on CPU
centered tasks but kernel-centered tasks.

One of this more interesting subsystem which fits the criteria is the network
subsystem. However, since we do not want to wait for IO, this should happen
locally.

So, create a network workload stress test script.

Signed-off-by: Anton Gres <dalo@linutronix.de>